requirements.txt
torch==2.2.0
transformers==4.38.0
faiss-cpu==1.7.4
numpy==1.26.0
imagehash==4.3.1
Pillow==10.2.0
pdf2image==1.17.0
opencv-python==4.9.0.80
PyQt6==6.6.0
huggingface-hub==0.20.3
facenet-pytorch==2.5.3
blake3==0.4.1